Major reorganization: moved to a subdirectory and using an inner class
