+++ samples/CapabilitiesList/README.md
