+++ samples/CapabilitiesList/adjustments4openSUSE.txt
