+++ samples/CapabilitiesList/build.sh
