+++ samples/CapabilitiesList/help.txt
