+++ samples/CapabilitiesList/systemjssMakefile
@@ -40,11 +40,9 @@ FILE=
 
 
 CLASSES = \
-	Capabilities.java \
-	ListerForAll.java
+	CapabilitiesList.java
 
-
-MAIN = ListerForAll
+MAIN = CapabilitiesList
 
 
 default: classes
