+++ samples/CapabilitiesList/useownnssdbBuild.sh
